pyproject.toml
[tool.poetry]
name = "fast-api-template"
version = "0.1.0"
description = "FastAPI - Backend Template"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.108.0"}
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pre-commit = "^3.6.0"
pytest = "^7.4.4"
faker = "^22.2.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.27.0"
boto3 = "^1.34.29"
python-multipart = "^0.0.6"
alembic = "^1.13.1"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[flake8]
ignore = ['E203', 'E266', 'E501', 'W503', 'F403', 'F401']
max-line-length = 79
max-complexity = 18
select = ['B','C','E','F','W','T4','B9']
